refactor: reestructuring of entity group and of group generally, add new entity group_memberships
