+++ farm-service/src/main/java/com/kernotec/farm/command/group/GroupCreateCmd.java
@@ -2,7 +2,6 @@ package com.kernotec.farm.command.group;
 
 import com.kernotec.core.command.AbstractTransactionalRequiredCommand;
 import com.kernotec.farm.jpa.entity.Group;
-import com.kernotec.farm.jpa.enums.GroupActionEnum;
 import com.kernotec.farm.jpa.service.GroupService;
 import jakarta.validation.constraints.NotNull;
 import java.util.UUID;
@@ -24,10 +23,7 @@ public class GroupCreateCmd extends
         Group group = new Group();
 
         group.setName(request.getName());
-        group.setAction(request.getAction());
-        group.setRegionId(request.getRegionId());
-        group.setActivityId(request.getActivityId());
-        group.setActivityTypeId(request.getActivityTypeId());
+        group.setDescription(request.getDescription());
 
         group = groupService.save(group);
         return group.getId();
@@ -40,14 +36,6 @@ public class GroupCreateCmd extends
         @NotNull
         private final String name;
         @NotNull
-        private final GroupActionEnum action;
-        @NotNull
-        private final UUID requestStateId;
-        @NotNull
-        private final UUID regionId;
-        @NotNull
-        private final UUID activityId;
-        @NotNull
-        private final UUID activityTypeId;
+        private final String description;
     }
 }
