+++ farm-service/src/main/java/com/kernotec/farm/rest/dto/response/activity/ActivityResponse.java
@@ -3,14 +3,14 @@ package com.kernotec.farm.rest.dto.response.activity;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import com.kernotec.core.rest.dto.response.data.EntityResponse;
-import com.kernotec.farm.rest.dto.response.PublishingResponse;
 import com.kernotec.farm.rest.dto.response.account.AccountForActivityResponse;
 import com.kernotec.farm.rest.dto.response.activity.type.ActivityTypeForActivityResponse;
-import com.kernotec.farm.rest.dto.response.comment.CommentResponse;
+import com.kernotec.farm.rest.dto.response.comment.CommentForActivityResponse;
 import com.kernotec.farm.rest.dto.response.follow.FollowResponse;
 import com.kernotec.farm.rest.dto.response.friend.FriendForActivityResponse;
-import com.kernotec.farm.rest.dto.response.group.GroupResponse;
-import com.kernotec.farm.rest.dto.response.reaction.ReactionResponse;
+import com.kernotec.farm.rest.dto.response.group.membership.GroupMembershipForActivityResponse;
+import com.kernotec.farm.rest.dto.response.publishing.PublishingForActivityResponse;
+import com.kernotec.farm.rest.dto.response.reaction.ReactionForActivityResponse;
 import java.time.ZonedDateTime;
 import java.util.Set;
 import java.util.UUID;
@@ -31,10 +31,10 @@ public class ActivityResponse extends EntityResponse {
     private UUID activityTypeId;
     private ActivityTypeForActivityResponse activityType;
 
-    private Set<PublishingResponse> publishings;
-    private Set<ReactionResponse> reactions;
-    private Set<CommentResponse> comments;
-    private Set<GroupResponse> groups;
+    private Set<PublishingForActivityResponse> publishings;
+    private Set<ReactionForActivityResponse> reactions;
+    private Set<CommentForActivityResponse> comments;
+    private Set<GroupMembershipForActivityResponse> groupMemberships;
     private Set<FriendForActivityResponse> friends;
     private Set<FollowResponse> follows;
 }
