+++ farm-service/src/main/java/com/kernotec/farm/rest/dto/response/comment/CommentForActivityResponse.java
@@ -1,10 +1,9 @@
-package com.kernotec.farm.rest.dto.response;
+package com.kernotec.farm.rest.dto.response.comment;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import com.kernotec.core.rest.dto.response.data.EntityResponse;
 import com.kernotec.farm.rest.dto.response.publishing.context.PublishingContextResponse;
-import com.kernotec.farm.rest.dto.response.publishing.type.PublishingTypeResponse;
 import java.util.UUID;
 import lombok.Getter;
 import lombok.Setter;
@@ -12,12 +11,10 @@ import lombok.Setter;
 @Getter
 @Setter
 @JsonInclude(Include.NON_NULL)
-public class PublishingResponse extends EntityResponse {
+public class CommentForActivityResponse extends EntityResponse {
 
-    private String description;
-
-    private UUID publishingTypeId;
-    private PublishingTypeResponse publishingType;
+    private String comment;
+    private Boolean isAgreeComment;
 
     private UUID publishingContextId;
     private PublishingContextResponse publishingContext;
