+++ farm-service/src/main/java/com/kernotec/farm/rest/dto/response/group/GroupResponse.java
@@ -3,8 +3,6 @@ package com.kernotec.farm.rest.dto.response.group;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import com.kernotec.core.rest.dto.response.data.EntityResponse;
-import com.kernotec.farm.rest.dto.response.region.RegionResponse;
-import java.util.UUID;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -14,8 +12,5 @@ import lombok.Setter;
 public class GroupResponse extends EntityResponse {
 
     private String name;
-    private Boolean isJoin;
-
-    private UUID regionId;
-    private RegionResponse region;
+    private String description;
 }
