+++ farm-service/src/main/java/com/kernotec/farm/rest/dto/response/group/membership/GroupMembershipResponse.java
@@ -1,11 +1,15 @@
-package com.kernotec.farm.rest.dto.response.comment;
+package com.kernotec.farm.rest.dto.response.group.membership;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import com.kernotec.core.rest.dto.response.data.EntityResponse;
+import com.kernotec.farm.jpa.enums.GroupActionEnum;
 import com.kernotec.farm.rest.dto.response.activity.ActivityFlatResponse;
 import com.kernotec.farm.rest.dto.response.activity.type.ActivityTypeForActivityResponse;
+import com.kernotec.farm.rest.dto.response.group.GroupResponse;
 import com.kernotec.farm.rest.dto.response.publishing.context.PublishingContextResponse;
+import com.kernotec.farm.rest.dto.response.region.RegionResponse;
+import com.kernotec.farm.rest.dto.response.request.state.RequestStateResponse;
 import java.util.UUID;
 import lombok.Getter;
 import lombok.Setter;
@@ -13,13 +17,21 @@ import lombok.Setter;
 @Getter
 @Setter
 @JsonInclude(Include.NON_NULL)
-public class CommentResponse extends EntityResponse {
+public class GroupMembershipResponse extends EntityResponse {
 
-    private String comment;
-    private Boolean isAgreeComment;
+    private GroupActionEnum action;
+
+    private UUID groupId;
+    private GroupResponse group;
+
+    private UUID regionId;
+    private RegionResponse region;
 
     private UUID publishingContextId;
     private PublishingContextResponse publishingContext;
+
+    private UUID requestStateId;
+    private RequestStateResponse requestState;
 
     private UUID activityId;
     private ActivityFlatResponse activity;
