+++ farm-service/src/main/java/com/kernotec/farm/rest/dto/response/publishing/PublishingResponse.java
@@ -1,11 +1,12 @@
-package com.kernotec.farm.rest.dto.response.reaction;
+package com.kernotec.farm.rest.dto.response.publishing;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import com.kernotec.core.rest.dto.response.data.EntityResponse;
 import com.kernotec.farm.rest.dto.response.activity.ActivityFlatResponse;
 import com.kernotec.farm.rest.dto.response.activity.type.ActivityTypeForActivityResponse;
-import com.kernotec.farm.rest.dto.response.reaction.type.ReactionTypeResponse;
+import com.kernotec.farm.rest.dto.response.publishing.context.PublishingContextResponse;
+import com.kernotec.farm.rest.dto.response.publishing.type.PublishingTypeResponse;
 import java.util.UUID;
 import lombok.Getter;
 import lombok.Setter;
@@ -13,10 +14,15 @@ import lombok.Setter;
 @Getter
 @Setter
 @JsonInclude(Include.NON_NULL)
-public class ReactionResponse extends EntityResponse {
+public class PublishingResponse extends EntityResponse {
 
-    private UUID reactionTypeId;
-    private ReactionTypeResponse reactionType;
+    private String description;
+
+    private UUID publishingTypeId;
+    private PublishingTypeResponse publishingType;
+
+    private UUID publishingContextId;
+    private PublishingContextResponse publishingContext;
 
     private UUID activityId;
     private ActivityFlatResponse activity;
