+++ farm-service/src/main/java/com/kernotec/farm/rest/dto/response/reaction/ReactionForActivityResponse.java
@@ -3,8 +3,6 @@ package com.kernotec.farm.rest.dto.response.reaction;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import com.kernotec.core.rest.dto.response.data.EntityResponse;
-import com.kernotec.farm.rest.dto.response.activity.ActivityFlatResponse;
-import com.kernotec.farm.rest.dto.response.activity.type.ActivityTypeForActivityResponse;
 import com.kernotec.farm.rest.dto.response.reaction.type.ReactionTypeResponse;
 import java.util.UUID;
 import lombok.Getter;
@@ -13,14 +11,8 @@ import lombok.Setter;
 @Getter
 @Setter
 @JsonInclude(Include.NON_NULL)
-public class ReactionResponse extends EntityResponse {
+public class ReactionForActivityResponse extends EntityResponse {
 
     private UUID reactionTypeId;
     private ReactionTypeResponse reactionType;
-
-    private UUID activityId;
-    private ActivityFlatResponse activity;
-
-    private UUID activityTypeId;
-    private ActivityTypeForActivityResponse activityType;
 }
