+++ farm-service/src/main/resources/db/changelog/2025/09/02-01-changelog-115.xml
@@ -1,0 +1,115 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<databaseChangeLog
+  xmlns="http://www.liquibase.org/xml/ns/dbchangelog"
+  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+  xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog
+                      http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-4.20.xsd"
+  objectQuotingStrategy="QUOTE_ONLY_RESERVED_WORDS">
+  <changeSet id="1756844398701-9" author="mmoya@PragmaInvest.com">
+    <dropForeignKeyConstraint baseTableName="groups" constraintName="fk_groups_on_activity"/>
+  </changeSet>
+  <changeSet id="1756844398701-10" author="mmoya@PragmaInvest.com">
+    <dropForeignKeyConstraint baseTableName="groups" constraintName="fk_groups_on_activity_type"/>
+  </changeSet>
+  <changeSet id="1756844398701-11" author="mmoya@PragmaInvest.com">
+    <dropForeignKeyConstraint baseTableName="groups"
+      constraintName="fk_groups_on_publishing_context"/>
+  </changeSet>
+  <changeSet id="1756844398701-12" author="mmoya@PragmaInvest.com">
+    <dropForeignKeyConstraint baseTableName="groups" constraintName="fk_groups_on_region"/>
+  </changeSet>
+  <changeSet id="1756844398701-13" author="mmoya@PragmaInvest.com">
+    <dropForeignKeyConstraint baseTableName="groups" constraintName="fk_groups_on_request_state"/>
+  </changeSet>
+  <changeSet id="1756844398701-1" author="mmoya@PragmaInvest.com">
+    <createTable tableName="group_memberships">
+      <column defaultValueComputed="uuid_generate_v4()" name="id" type="uuid">
+        <constraints nullable="false" primaryKey="true" primaryKeyName="pk_group_memberships"/>
+      </column>
+      <column defaultValueComputed="now()" name="created_at" type="timestamp">
+        <constraints nullable="false"/>
+      </column>
+      <column name="created_by" type="VARCHAR(255)"/>
+      <column defaultValueComputed="now()" name="updated_at" type="timestamp">
+        <constraints nullable="false"/>
+      </column>
+      <column name="updated_by" type="VARCHAR(255)"/>
+      <column defaultValueNumeric="0" name="version" type="bigint">
+        <constraints nullable="false"/>
+      </column>
+      <column defaultValueBoolean="false" name="deleted" type="boolean">
+        <constraints nullable="false"/>
+      </column>
+      <column name="action" type="VARCHAR(255)">
+        <constraints nullable="false"/>
+      </column>
+      <column name="group_id" type="UUID">
+        <constraints nullable="false"/>
+      </column>
+      <column name="region_id" type="UUID">
+        <constraints nullable="false"/>
+      </column>
+      <column name="publishing_context_id" type="UUID">
+        <constraints nullable="false"/>
+      </column>
+      <column name="request_state_id" type="UUID">
+        <constraints nullable="false"/>
+      </column>
+      <column name="activity_id" type="UUID">
+        <constraints nullable="false"/>
+      </column>
+      <column name="activity_type_id" type="UUID">
+        <constraints nullable="false"/>
+      </column>
+    </createTable>
+  </changeSet>
+  <changeSet id="1756844398701-2" author="mmoya@PragmaInvest.com">
+    <addColumn tableName="groups">
+      <column name="description" type="VARCHAR(500)"/>
+    </addColumn>
+  </changeSet>
+  <changeSet id="1756844398701-3" author="mmoya@PragmaInvest.com">
+    <addForeignKeyConstraint baseColumnNames="activity_id" baseTableName="group_memberships"
+      constraintName="FK_GROUP_MEMBERSHIPS_ON_ACTIVITY" referencedColumnNames="id"
+      referencedTableName="activities"/>
+  </changeSet>
+  <changeSet id="1756844398701-4" author="mmoya@PragmaInvest.com">
+    <addForeignKeyConstraint baseColumnNames="activity_type_id" baseTableName="group_memberships"
+      constraintName="FK_GROUP_MEMBERSHIPS_ON_ACTIVITY_TYPE" referencedColumnNames="id"
+      referencedTableName="activity_types"/>
+  </changeSet>
+  <changeSet id="1756844398701-5" author="mmoya@PragmaInvest.com">
+    <addForeignKeyConstraint baseColumnNames="group_id" baseTableName="group_memberships"
+      constraintName="FK_GROUP_MEMBERSHIPS_ON_GROUP" referencedColumnNames="id"
+      referencedTableName="groups"/>
+  </changeSet>
+  <changeSet id="1756844398701-6" author="mmoya@PragmaInvest.com">
+    <addForeignKeyConstraint baseColumnNames="publishing_context_id"
+      baseTableName="group_memberships" constraintName="FK_GROUP_MEMBERSHIPS_ON_PUBLISHING_CONTEXT"
+      referencedColumnNames="id" referencedTableName="publishing_contexts"/>
+  </changeSet>
+  <changeSet id="1756844398701-7" author="mmoya@PragmaInvest.com">
+    <addForeignKeyConstraint baseColumnNames="region_id" baseTableName="group_memberships"
+      constraintName="FK_GROUP_MEMBERSHIPS_ON_REGION" referencedColumnNames="id"
+      referencedTableName="regions"/>
+  </changeSet>
+  <changeSet id="1756844398701-8" author="mmoya@PragmaInvest.com">
+    <addForeignKeyConstraint baseColumnNames="request_state_id" baseTableName="group_memberships"
+      constraintName="FK_GROUP_MEMBERSHIPS_ON_REQUEST_STATE" referencedColumnNames="id"
+      referencedTableName="request_states"/>
+  </changeSet>
+  <changeSet id="1756844398701-14" author="mmoya@PragmaInvest.com">
+    <dropColumn columnName="action" tableName="groups"/>
+
+    <dropColumn columnName="activity_id" tableName="groups"/>
+
+    <dropColumn columnName="activity_type_id" tableName="groups"/>
+
+    <dropColumn columnName="publishing_context_id" tableName="groups"/>
+
+    <dropColumn columnName="region_id" tableName="groups"/>
+
+    <dropColumn columnName="request_state_id" tableName="groups"/>
+  </changeSet>
+
+</databaseChangeLog>
