+++ farm-service/src/main/resources/db/changelog/db.changelog-master.xml
@@ -85,8 +85,9 @@
   </changeSet>
 
   <include file="db/changelog/2025/09/02-01-changelog-114.xml"/>
+  <include file="db/changelog/2025/09/02-01-changelog-115.xml"/>
 
-    <!-- social network seeder update 3 - add colors -->
+  <!-- social network seeder update 3 - add colors -->
     <changeSet id="1756825981012-1" author="Maur025">
       <sqlFile path="data/social/network/social_network_seeder_update3.sql" relativeToChangelogFile="true"/>
     </changeSet>
